refactor(DUP): move `resend_all_properties` to dedicated device module
Move `resend_all_properties/1` away from `impl.ex` and test separately
